Expose host CPU to qemu and avoid missing features required by glibc

+++ launch-nodes.sh
@@ -72,6 +72,7 @@ for i in "${!hosts[@]}"; do
     qemu-system-x86_64 \
       -m 4G \
       -smp 2 \
+      -cpu host \
       -device virtio-net-pci,netdev=net0,mac=${hosts[$i]} \
       -netdev bridge,id=net0,br=br0 \
       -drive file=./image_$i.qcow2,if=virtio,cache=writeback,discard=ignore,format=qcow2 \
@@ -84,18 +85,17 @@ for i in "${!hosts[@]}"; do
 done
 
 # get k3s.yaml to use with `k9s --kubeconfig k3s.yaml` or `kubectl --kubeconfig k3s.yaml ...`
-while true; do
-  K3S_CONFIG=$(ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i ./k3s_cluster_key root@10.0.0.10 'cat /etc/rancher/k3s/k3s.yaml 2>/dev/null' || true)
-  if [ -n "$K3S_CONFIG" ]; then
-    echo "Got k3s config"
-    break
-  fi
-  echo "Waiting for k3s config..." || true
-  sleep 5
-done
-
-printf "%s" "$K3S_CONFIG" | sed 's/127\.0\.0\.1/10.0.0.10/g' > k3s.yaml
-
-chmod o+r k3s_cluster_key # in case this script is run as root but we want to use the key
+if [ ! -f "k3s.yaml" ] || [ "$RECREATE" = true ]; then
+  while true; do
+    K3S_CONFIG=$(ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -i ./k3s_cluster_key root@10.0.0.10 'cat /etc/rancher/k3s/k3s.yaml 2>/dev/null' || true)
+    if [ -n "$K3S_CONFIG" ]; then
+      echo "Got k3s config"
+      break
+    fi
+    echo "Waiting for k3s config..." || true
+    sleep 5
+  done
+  printf "%s" "$K3S_CONFIG" | sed 's/127\.0\.0\.1/10.0.0.10/g' > k3s.yaml
+fi
 
 wait
